1 Пакеты работы с переменными окружения и флагами перемещены в пакет конфигурации
